requirements.txt
tk
requests
Pillow
